backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic",
    "numpy",
    "ollama",
    "faiss-cpu",
    "ipykernel>=7.2.0",
    "fastapi",
    "uvicorn[standard]",
    "openai",
    "google-genai",
    "python-dotenv",
]
